Add refactory tool and update the refactory benchmark
